tModLoader1.4.4移植准备

+++ Content/NPCs/StateMachine/SMNPC.cs
@@ -15,7 +15,7 @@
         public int State
         {
             get { return (int)NPC.ai[0]; }
-            set { NPC.ai[0] = (int)value; }
+            set { NPC.ai[0] = value; }
         }
         /// <summary>
         /// 计时器
